app/pyproject.toml
[tool.poetry]
name = "NHSX"
version = "0.1.0"
description = "NHSX Wagtail site"
authors = ["andy@hactar.is"]

[tool.poetry.dependencies]
python = "^3.6"

wagtail = "2.8.1"

psycopg2 = "^2.8"
django-redis = "^4.10"
hiredis = "^1.0"
django-storages = "^1.7"
boto3 = "^1.9"
whitenoise = "^4.1"
envkey = "^1.2"
bugsnag = "^3.6"
wagtail-cache = "^1.0.0"
wagtailfontawesome = "^1.1"
django_extensions = "^2.2"
django-cacheops = "^4.2"
arrow = "^0.15.2"
wagtail-modelchooser = "^2.2"
consoler = "^0.1.2"
wagtail_cache = "^1.0.0"
python-memcached = "^1.59"
wagtail-nhsuk-frontend = "^0.4.1"
uwsgi = "^2.0.18"
django-assets = "^2.0"
django-autocomplete-light = "^3.5"
wagtail-2fa = "^1.4.2"
boto = "^2.49.0"
rollbar = "^0.15.0"
black = "^19.10b0"



[tool.poetry.dev-dependencies]
codacy-coverage = "*"
coverage = "*"
csscompressor = "*"
cssmin = "*"
factory-boy = "*"
jsmin = "*"
mixer = "*"
ipdb = "*"
pdbpp = "*"
pytest = "5.3.5"
pytest-cov = "*"
pytest-django = "*"
pytest-factoryboy = "*"
pytest-pythonpath = "*"
pytest-sugar = "*"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
